# Conflicts:
#	VERSION
#	pom.xml
#	pulsar-app/pom.xml
#	pulsar-app/pulsar-examples/pom.xml
#	pulsar-common/pom.xml
#	pulsar-distribution/pom.xml
#	pulsar-dom/pom.xml
#	pulsar-persist/pom.xml
#	pulsar-plugins/pom.xml
#	pulsar-plugins/pulsar-filter/pom.xml
#	pulsar-plugins/pulsar-index/pom.xml
#	pulsar-plugins/pulsar-parse/pom.xml
#	pulsar-plugins/pulsar-protocol/pom.xml
#	pulsar-plugins/pulsar-schedule/pom.xml
#	pulsar-plugins/pulsar-scoring/pom.xml
#	pulsar-ql-common/pom.xml
#	pulsar-ql-server/pom.xml
#	pulsar-rest/pom.xml
#	pulsar-skeleton/pom.xml
#	pulsar-spring-support/pulsar-beans/pom.xml
#	pulsar-third/pom.xml
#	pulsar-third/pulsar-boilerpipe/pom.xml
#	pulsar-third/pulsar-chrome-devtools/pom.xml
#	pulsar-third/pulsar-jsoup/pom.xml
#	pulsar-tools/pom.xml
#	pulsar-tools/pulsar-browser/pom.xml

+++ pulsar-tools/pulsar-browser/src/main/kotlin/ai/platon/pulsar/browser/driver/chrome/ChromeLauncher.kt
@@ -329,9 +329,9 @@ class ChromeLauncher(
                 throw ChromeProcessTimeoutException("Timeout to waiting for chrome to start")
             }
         } catch (e: InterruptedException) {
+            Thread.currentThread().interrupt()
+            log.error("Interrupted while waiting for devtools server, close it", e)
             close(readLineThread)
-            log.error("Interrupted while waiting for dev tools server", e)
-            throw RuntimeException("Interrupted while waiting for dev tools server", e)
         }
         return port
     }
@@ -339,24 +339,39 @@ class ChromeLauncher(
     private fun close(thread: Thread) {
         try {
             thread.join(config.threadWaitTime.toMillis())
-        } catch (ignored: InterruptedException) {}
+        } catch (e: InterruptedException) {
+            Thread.currentThread().interrupt()
+        }
     }
 
     private fun prepareUserDataDir() {
-        val lock = AppPaths.BROWSER_TMP_DIR_LOCK
         val prototypeUserDataDir = AppPaths.CHROME_DATA_DIR_PROTOTYPE
+        if (userDataDir == prototypeUserDataDir || userDataDir.toString().contains("/default/")) {
+            log.info("Running chrome with prototype/default data dir, no cleaning | {}", userDataDir)
+            return
+        }
+
+        val lock = AppPaths.BROWSER_TMP_DIR_LOCK
         if (Files.exists(prototypeUserDataDir.resolve("Default"))) {
             FileChannel.open(lock, StandardOpenOption.APPEND).use {
                 it.lock()
 
                 if (!Files.exists(userDataDir.resolve("Default"))) {
                     log.info("User data dir does not exist, copy from prototype | {} <- {}", userDataDir, prototypeUserDataDir)
+                    // remove dead symbolic links
+                    Files.list(prototypeUserDataDir).filter { Files.isSymbolicLink(it) && !Files.exists(it) }.forEach { Files.delete(it) }
                     FileUtils.copyDirectory(prototypeUserDataDir.toFile(), userDataDir.toFile())
                 } else {
                     Files.deleteIfExists(userDataDir.resolve("Default/Cookies"))
-                    FileUtils.deleteDirectory(userDataDir.resolve("Default/Local Storage/leveldb").toFile())
+                    val leveldb = userDataDir.resolve("Default/Local Storage/leveldb")
+                    if (Files.exists(leveldb)) {
+                        FileUtils.deleteDirectory(leveldb.toFile())
+                    }
+
                     arrayOf("Default/Cookies", "Default/Local Storage/leveldb").forEach {
-                        Files.copy(prototypeUserDataDir.resolve(it), userDataDir.resolve(it), StandardCopyOption.REPLACE_EXISTING)
+                        val target = userDataDir.resolve(it)
+                        Files.createDirectories(target.parent)
+                        Files.copy(prototypeUserDataDir.resolve(it), target, StandardCopyOption.REPLACE_EXISTING)
                     }
                 }
             }
